+++ NitroxClient/GameLogic/HUD/Components/DenyOwnershipHand.cs
@@ -1,8 +1,6 @@
-﻿
+﻿using UnityEngine;
 
-using UnityEngine;
-
-namespace NitroxClient.GameLogic.HUD
+namespace NitroxClient.GameLogic.HUD.Components
 {
     public class DenyOwnershipHand : MonoBehaviour
     {
